Post comment bug found and fixed
