+++ components/CampsiteInfoComponent.js
@@ -1,53 +1,31 @@
 import React, { Component } from "react";
-import { Text, View, ScrollView, FlatList, Modal, Button, StyleSheet } from "react-native";
-import { Card, Icon, Rating, Input } from "react-native-elements";
+import {
+  Text,
+  View,
+  ScrollView,
+  FlatList,
+  Modal,
+  Button,
+  StyleSheet
+} from "react-native";
+import { Card, Icon, Input, Rating } from "react-native-elements";
 import { connect } from "react-redux";
 import { baseUrl } from "../shared/baseUrl";
 import { postFavorite, postComment } from "../redux/ActionCreators";
 
-
-const mapStateToProps = (state) => {
+const mapStateToProps = state => {
   return {
     campsites: state.campsites,
     comments: state.comments,
-    favorites: state.favorites,
+    favorites: state.favorites
   };
 };
 
 const mapDispatchToProps = {
-  postFavorite: (campsiteId) => postFavorite(campsiteId),
+  postFavorite: campsiteId => postFavorite(campsiteId),
   postComment: (campsiteId, rating, author, text) =>
-    postComment(campsiteId, rating, author, text),
+    postComment(campsiteId, rating, author, text)
 };
-
-function RenderComments({ comments }) {
-  const renderCommentItem = ({ item }) => {
-    return (
-      <View style={{ margin: 10 }}>
-        <Text style={{ fontSize: 14 }}>{item.text}</Text>
-        <Rating
-          startingValue={item.rating}
-          readonly
-          imageSize={10}
-          style={{ alignItems: "flex-start", paddingVertical: "5%" }}
-        />
-        <Text
-          style={{ fontSize: 12 }}
-        >{`-- ${item.author}, ${item.date}`}</Text>
-      </View>
-    );
-  };
-
-  return (
-    <Card title="Comments">
-      <FlatList
-        data={comments}
-        renderItem={renderCommentItem}
-        keyExtractor={(item) => item.id.toString()}
-      />
-    </Card>
-  );
-}
 
 function RenderCampsite(props) {
   const { campsite } = props;
@@ -76,6 +54,7 @@ function RenderCampsite(props) {
             name={"pencil"}
             type="font-awesome"
             color="#5637DD"
+            style={styles.cardItem}
             raised
             reverse
             onPress={() => props.onShowModal()}
@@ -86,15 +65,46 @@ function RenderCampsite(props) {
   }
   return <View />;
 }
+
+function RenderComments({ comments }) {
+  const renderCommentItem = ({ item }) => {
+    return (
+      <View style={{ margin: 10 }}>
+        <Text style={{ fontSize: 14 }}>{item.text}</Text>
+        <Rating
+          type="star"
+          fractions={0}
+          startingValue={item.rating}
+          imageSize={10}
+          readonly
+          style={{ alignItems: "flex-start", paddingVertical: "5%" }}
+        />
+        <Text
+          style={{ fontSize: 12 }}
+        >{`-- ${item.author}, ${item.date}`}</Text>
+      </View>
+    );
+  };
+
+  return (
+    <Card title="Comments">
+      <FlatList
+        data={comments}
+        renderItem={renderCommentItem}
+        keyExtractor={item => item.id.toString()}
+      />
+    </Card>
+  );
+}
+
 class CampsiteInfo extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      favorite: false,
-      showModal: false,
       rating: 5,
       author: "",
       text: "",
+      showModal: false
     };
   }
 
@@ -103,35 +113,39 @@ class CampsiteInfo extends Component {
   }
 
   handleComment(campsiteId) {
-    this.props.postComment(campsiteId, this.state.rating, this.state.author, this.state.text);
     this.toggleModal();
+    this.props.postComment(
+      campsiteId,
+      this.state.rating,
+      this.state.author,
+      this.state.text
+    );
   }
 
   resetForm() {
     this.setState({
-      favorite: false,
-      showModal: false,
       rating: 5,
       author: "",
       text: "",
-    })
+      showModal: false
+    });
   }
-
-  static navigationOptions = {
-    title: "Campsite Information",
-  };
 
   markFavorite(campsiteId) {
     this.props.postFavorite(campsiteId);
   }
 
+  static navigationOptions = {
+    title: "Campsite Information"
+  };
+
   render() {
     const campsiteId = this.props.navigation.getParam("campsiteId");
     const campsite = this.props.campsites.campsites.filter(
-      (campsite) => campsite.id === campsiteId
+      campsite => campsite.id === campsiteId
     )[0];
     const comments = this.props.comments.comments.filter(
-      (comment) => comment.campsiteId === campsiteId
+      comment => comment.campsiteId === campsiteId
     );
     return (
       <ScrollView>
@@ -151,43 +165,49 @@ class CampsiteInfo extends Component {
           <View style={styles.modal}>
             <Rating
               showRating
+              type="star"
+              fractions={0}
               startingValue={this.state.rating}
               imageSize={40}
-              onFinishRating={(rating) => this.setState({ rating: rating })}
+              onFinishRating={rating => this.setState({ rating: rating })}
               style={{ paddingVertical: 10 }}
             />
             <Input
               placeholder="Author"
               leftIcon={{ type: "font-awesome", name: "user-o" }}
               leftIconContainerStyle={{ paddingRight: 10 }}
-              onChangeText={(author) => this.setState({ author: author })}
+              onChangeText={author => {
+                this.setState({ author: author });
+              }}
               value={this.state.author}
             />
             <Input
               placeholder="Comment"
               leftIcon={{ type: "font-awesome", name: "comment-o" }}
               leftIconContainerStyle={{ paddingRight: 10 }}
-              onChangeText={(comment) => this.setState({ text: comment })}
+              onChangeText={text => {
+                this.setState({ text: text });
+              }}
               value={this.state.text}
             />
             <View style={{ margin: 10 }}>
-              <Button 
-              title={"Submit"}
-              color="#5637DD"
-              onPress={(campsiteId) => {
-                this.handleComment(campsiteId);
-                this.resetForm();
-              }}
+              <Button
+                onPress={() => {
+                  this.handleComment(campsiteId);
+                  this.resetForm();
+                }}
+                color="#5637DD"
+                title="Submit"
               />
             </View>
             <View style={{ margin: 10 }}>
               <Button
                 onPress={() => {
-                  this.resetForm();
                   this.toggleModal();
+                  this.resetForm();
                 }}
                 color="#808080"
-                title={"Cancel"}
+                title="Cancel"
               />
             </View>
           </View>
@@ -203,12 +223,15 @@ const styles = StyleSheet.create({
     justifyContent: "center",
     flex: 1,
     flexDirection: "row",
-    margin: 20,
+    margin: 20
+  },
+  cardItem: {
+    flex: 1,
+    margin: 10
   },
   modal: {
     justifyContent: "center",
-    margin: 20,
-
+    margin: 20
   }
 });
